fix(ci): repair GitHub Actions workflow for CI for Windows

+++ build.ps1
@@ -1,5 +1,8 @@
-Remove-Item -Recurse -Force build
-New-Item -ItemType Directory -Path build
+if (Test-Path build) {
+    Remove-Item -Recurse -Force build
+}
+
+New-Item -ItemType Directory -Path build -ErrorAction SilentlyContinue | Out-Null
 Set-Location build
 cmake ..
 cmake --build .
